Небольшой рефакторинг

+++ GameLogic/Board.cs
@@ -69,9 +69,9 @@ public class Board
 
     public IEnumerable<Position> PiecePositions()
     {
-        for(int r = 0; r < 8; r++)
+        for(ushort r = 0; r < 8; r++)
         {
-            for (int c = 0; c < 8; c++)
+            for (ushort c = 0; c < 8; c++)
             {
                 Position pos = new Position(r, c);
